about page and meet the team page swap plus changes
