+++ src/pages/TeamPage.jsx
@@ -1,145 +1,210 @@
-import calendarImage from "../images/calendar.png";
-import budgetImage from "../images/Budgets.png";
+import { useState, useEffect } from "react";
+import Mike from "../images/Mike.png";
+import Jose from "../images/Jose.png";
+import Weston from "../images/Weston.png";
+import Saul from "../images/Saul.png";
+import Jorge from "../images/Jorge.png";
+import Andrew from "../images/Andrew.png";
+import cashregister from "../images/cashregister.mp3";
+// https://pixabay.com/sound-effects/search/cash/
+// Cash Register Purchase MP3 Royalty Free, Free Download
+const AboutPage = () => {
+  const [cylinderRotation, setCylinderRotation] = useState(0);
+  const [holeRotation, setHoleRotation] = useState(0);
+  const [isPlaying, setIsPlaying] = useState(false);
 
-import { Divider } from "@tremor/react";
+  useEffect(() => {
+    if (isPlaying) {
+      const audio = new Audio(cashregister);
+      audio.play();
+      audio.addEventListener('ended', () => {
+        setIsPlaying(false);
+      });
+    }
+  }, [isPlaying]);
 
-const TeamPage = () => {
+  const handleRotate = () => {
+    setIsPlaying(true);
+    setCylinderRotation((prevRotation) => prevRotation + 60); 
+    setHoleRotation((prevRotation) => prevRotation); 
+  };
+
+  const getHoleStyle = (initialAngle, translateX) => ({
+    transform: `rotate(${
+      holeRotation + initialAngle
+    }deg) translateX(${translateX}px) rotate(-${
+      holeRotation + initialAngle + cylinderRotation
+    }deg)`,
+  });
+
+  const initialAngles = {
+    Jose: -90,
+    Mike: -30,
+    Weston: 30,
+    Saul: 90,
+    Jorge: 150,
+    Andrew: 210,
+  };
+
+  const findNameWithAngle = () => {
+    for (const name in initialAngles) {
+      const angle = (initialAngles[name] + cylinderRotation) % 360;
+      if (angle === 270) {
+        return name;
+      }
+    }
+    return null; // Return null if no name matches the angle
+  };
+
+  const nameWithAngle270 = findNameWithAngle();
+
+  console.log(holeRotation);
   return (
     <>
-      <section className='rounded-xl shadow-xl shadow-slate-300/70 bg-slate-400 p-4 mx-5 my-4'>
-        <div className='grid my-2 max-w-screen-3xl rounded-xl  px-4 py-8 mx-auto lg:gap-8 xl:gap-0 lg:py-16 lg:grid-cols-12'>
-          <div className='mr-auto rounded-xl p-1 place-self-center lg:col-span-7'>
-            <h1 className='max-w-2xl p-2 mb-4 text-4xl font-extrabold tracking-tight leading-none md:text-5xl xl:text-6xl '>
-              Visualize your finances with our interactive calendar
-            </h1>
-            <p className='max-w-2xl mb-6 font-light text-slate-300 lg:mb-8 md:text-lg lg:text-xl dark:text-gray-400'>
-              From upcoming recurring payments or income, to individual
-              transcations all in one place to quickly grasp upcoming or past
-              events.
-            </p>
-            <a
-              href='#'
-              className='inline-flex items-center justify-center px-5 py-3 mr-3 text-base font-medium text-center text-slate-100 rounded-lg bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900'
-            >
-              Get started
-              <svg
-                className='w-5 h-5 ml-2 -mr-1'
-                fill='currentColor'
-                viewBox='0 0 20 20'
-                xmlns='http://www.w3.org/2000/svg'
-              >
-                <path d='M10.293 3.293a1 1 0 011.414 0l6 6a1 1 0 010 1.414l-6 6a1 1 0 01-1.414-1.414L14.586 11H3a1 1 0 110-2h11.586l-4.293-4.293a1 1 0 010-1.414z'></path>
-              </svg>
-            </a>
-            <a
-              href='#'
-              className='inline-flex bg-slate-200 items-center justify-center px-5 py-3 text-base font-medium text-center text-slate-900 border border-slate-300 rounded-lg hover:bg-gray-100 focus:ring-4 focus:ring-gray-100'
-            >
-              Sign Up Today!
-            </a>
+      <div className="about-page-container">
+        <p className="quote-style">
+          &quot;TBD: Helping you hit your financial targets.&quot;
+        </p>
+        {nameWithAngle270 && (
+          <>
+            {nameWithAngle270 === "Jose" && (
+              <p className="bio">Jose Sandoval, Army Veteran turned Full-Stack Software Engineer, worked on front-end development and UI design, contributing broadly to ensure the project's success. From Cars & Coffee to Code & Coffee, you'll find me meeting people and sharing passions!</p>
+            )}
+            {nameWithAngle270 === "Mike" && (
+              <p className="bio">Mike Durell, Navy Intelligence Veteran turned Full-Stack Software Engineer, worked on the animations and styling for the home, login/signup, and about us pages.</p>
+            )}
+            {nameWithAngle270 === "Weston" && (
+              <p className="bio">Weston Kelly, Marine Corps Veteran turned Full-Stack Software Engineer, worked on authentication, demo video, chart and calendar functionality for the application.</p>
+            )}
+            {nameWithAngle270 === "Saul" && (
+              <p className="bio">Saul Alvarez, Navy Engineering Veteran turned Full-Stack Software Engineer, worked on the database, data entry, table design and line chart functionality.</p>
+            )}
+            {nameWithAngle270 === "Jorge" && (
+              <p className="bio">Jorge L. Ayala, Marine Corps Veteran turned Full-Stack Software Engineer, worked on data manipulation, dynamic charts, and interactive calendars.</p>
+            )}
+            {nameWithAngle270 === "Andrew" && (
+              <p className="bio">Andrew Dew, Special Forces Veteran turned Full-Stack Software Engineer, worked on the database, data manipulation, and table design.</p>
+            )}
+          </>
+        )}
+    
+        <div
+          className="cylinder"
+          onClick={handleRotate}
+          style={{ transform: `rotate(${cylinderRotation}deg)` }}
+        >
+          <div
+            className="hole"
+            style={{ ...getHoleStyle(-87.5, 143)}}
+          >
+            {nameWithAngle270 === "Jose" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Jose} alt="Jose" />
           </div>
-          <div className='hidden rounded-2xl shadow-xl shadow-slate-200/70 lg:mt-0 lg:col-span-5 lg:flex'>
-            <img
-              className='rounded-2xl w-full h-full'
-              src={calendarImage}
-              alt='mockup'
-            />
+          <div
+            className="hole"
+            style={{ ...getHoleStyle(-29.5, 149)}}
+          >
+            {nameWithAngle270 === "Mike" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Mike} alt="Mike" />
+          </div>
+          <div
+            className="hole"
+            style={{ ...getHoleStyle(26, 150)}}
+          >
+            {nameWithAngle270 === "Weston" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Weston} alt="Weston" />
+          </div>
+          <div
+            className="hole picture-adjust4"
+            style={{ ...getHoleStyle(84, 143)}}
+          >
+            {nameWithAngle270 === "Saul" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Saul} alt="Saul" />
+          </div>
+          <div
+            className="hole picture-adjust5"
+            style={{ ...getHoleStyle(147, 137)}}
+          >
+            {nameWithAngle270 === "Jorge" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Jorge} alt="Jorge" />
+          </div>
+          <div
+            className="hole picture-adjust6"
+            style={{ ...getHoleStyle(211, 138)}}
+          >
+            {nameWithAngle270 === "Andrew" && (
+              <div
+                style={{
+                  position: "absolute",
+                  top: 0,
+                  left: 0,
+                  width: "100%",
+                  height: "100%",
+                  backgroundColor: "rgba(0, 255, 0, 0.5)",
+                }}
+              />
+            )}
+            <img src={Andrew} alt="Andrew" />
           </div>
         </div>
-        <div className='py-8'>
-          <Divider> ◦◦◦⦿◦◦◦</Divider>
-        </div>
-
-        <div className='flex justify-center'>
-          <img
-            className='rounded-2xl w-2/3'
-            src={budgetImage}
-            alt='budget table'
-          />
-        </div>
-        <div className='flex justify-center w-full items-center content-center'>
-          <div className='flex justify-center items-center'>
-            <div className='grid max-w-screen-2xl rounded-xl mx-auto lg:gap- xl:gap-0 lg:py-16 '>
-              <div className='mr-auto rounded-xl p-1 place-self-center lg:col-span-7'>
-                <h1 className='max-w-2xl  text-4xl font-extrabold tracking-tight leading-none md:text-5xl xl:text-6xl '>
-                  Budget Tracking at an instance
-                </h1>
-                <p className='max-w-2xl mb-6 font-light text-slate-300 lg:mb-8 md:text-lg lg:text-xl dark:text-gray-400'>
-                  Never lose track of your budget data again, with our budgets
-                  feature, you can find yourself in charge and prepared!
-                </p>
-                <a
-                  href='#'
-                  className='inline-flex items-center justify-center px-5 py-3 mr-3 text-base font-medium text-center text-slate-100 rounded-lg bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900'
-                >
-                  Get started
-                  <svg
-                    className='w-5 h-5 ml-2 -mr-1'
-                    fill='currentColor'
-                    viewBox='0 0 20 20'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path d='M10.293 3.293a1 1 0 011.414 0l6 6a1 1 0 010 1.414l-6 6a1 1 0 01-1.414-1.414L14.586 11H3a1 1 0 110-2h11.586l-4.293-4.293a1 1 0 010-1.414z'></path>
-                  </svg>
-                </a>
-                <a
-                  href='#'
-                  className='inline-flex bg-slate-200 items-center justify-center px-5 py-3 text-base font-medium text-center text-slate-900 border border-slate-300 rounded-lg hover:bg-gray-100 focus:ring-4 focus:ring-gray-100'
-                >
-                  Sign Up Today!
-                </a>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className='py-8'>
-          <Divider> ◦◦◦⦿◦◦◦</Divider>
-        </div>
-        <div className='flex justify-center'>
-          <img
-            className='rounded-2xl w-2/3'
-            src={budgetImage}
-            alt='budget table'
-          />
-        </div>
-        <div className='flex justify-center w-full items-center content-center'>
-          <div className='flex justify-center items-center'>
-            <div className='grid max-w-screen-2xl rounded-xl mx-auto lg:gap- xl:gap-0 lg:py-16 '>
-              <div className='mr-auto rounded-xl p-1 place-self-center lg:col-span-7'>
-                <h1 className='max-w-2xl  text-4xl font-extrabold tracking-tight leading-none md:text-5xl xl:text-6xl '>
-                  Budget Tracking at an instance
-                </h1>
-                <p className='max-w-2xl mb-6 font-light text-slate-300 lg:mb-8 md:text-lg lg:text-xl dark:text-gray-400'>
-                  Never lose track of your budget data again, with our budgets
-                  feature, you can find yourself in charge and prepared!
-                </p>
-                <a
-                  href='#'
-                  className='inline-flex items-center justify-center px-5 py-3 mr-3 text-base font-medium text-center text-slate-100 rounded-lg bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900'
-                >
-                  Get started
-                  <svg
-                    className='w-5 h-5 ml-2 -mr-1'
-                    fill='currentColor'
-                    viewBox='0 0 20 20'
-                    xmlns='http://www.w3.org/2000/svg'
-                  >
-                    <path d='M10.293 3.293a1 1 0 011.414 0l6 6a1 1 0 010 1.414l-6 6a1 1 0 01-1.414-1.414L14.586 11H3a1 1 0 110-2h11.586l-4.293-4.293a1 1 0 010-1.414z'></path>
-                  </svg>
-                </a>
-                <a
-                  href='#'
-                  className='inline-flex bg-slate-200 items-center justify-center px-5 py-3 text-base font-medium text-center text-slate-900 border border-slate-300 rounded-lg hover:bg-gray-100 focus:ring-4 focus:ring-gray-100'
-                >
-                  Sign Up Today!
-                </a>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      </div>
     </>
   );
 };
 
-export default TeamPage;
+export default AboutPage;
